1. ID has Long type now instead of Int 2. Connection to Database was moved to the separate class ConnectionFactory 3. RowMapper field was created in JdbcUserDao class 4. RuntimeException is thrown in all methods of JdbcUserDao, UserRowMapper classes 5. All ResultSets are closed now 6. Methods’ names were simplified in UserService class; jdbcUserDao = new JdbcUserDao() was removed from each method. 7. RemoveUserServlet was created 8. EditUserServlet class: userService field was created. 9. “response.getWriter().println(PageGenerator.instance().getPage("…", pageVariables));” record was divided to three strings in all Servlets to simplify reading of it. 10. response.getWriter().println(page) was changed to response.getWriter().write(page) in all Servlets 11. method getPage (String filename) was added to PageGenerator class. PageGenerator is field in this class now. 12. Addeduser.html template: #else condition was removed 13. doPost method in AddNewUserServlet was reworked to provide users.html page. Addeduser.html template was removed. 14. title on all templates was changed 15. jetty versions in .pom file were updated 16. ApplicationProperties file was added. Classes Starter and ConnectionFactory use it now.

+++ src/main/java/com/boroday/userstore/entity/User.java
@@ -3,17 +3,17 @@ package com.boroday.userstore.entity;
 import java.time.LocalDate;
 
 public class User {
-    private int id;
+    private long id;
     private String firstName;
     private String lastName;
     private Double salary;
     private LocalDate dateOfBirth;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -49,7 +49,7 @@ public class User {
         this.dateOfBirth = dateOfBirth;
     }
 
-    public String toString(){
+    public String toString() {
         return "UserID = " + id + "; FirstName = " + firstName + "; LastName = " + lastName + "; Salary = " + salary + "; DateOfBirth = " + dateOfBirth + ".";
     }
 }
